refactor: move package to common position
