+++ tomcat/src/main/java/org/apache/coyote/http11/ProtocolVersion.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11;
 
 import java.util.Arrays;
