+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -9,6 +9,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import org.apache.coyote.http11.ProtocolVersion;
 
 public record HttpRequest(
         Method method,
