+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpResponse.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
-import org.apache.coyote.http11.request.ProtocolVersion;
+import org.apache.coyote.http11.ProtocolVersion;
 
 public record HttpResponse(ProtocolVersion protocolVersion, Status status,
                            Map<String, String> headers, Map<String, String> cookies, byte[] body) {
